@@ -1,7 +1,7 @@
 {
   "name": "react-grid",
   "version": "0.0.0",
-  "description": "Data grid fro React",
+  "description": "Data grid for React",
   "main": "lib/index.js",
   "scripts": {
     "test": "make test"
@@ -15,14 +15,42 @@
     "grid"
   ],
   "peerDependencies": {
-    "react": "^0.10.0"
+    "react": "^0.11.1"
   },
   "devDependencies": {
     "mocha": "^1.18.2",
-    "mochify": "^0.8.1",
-    "react": "^0.10.0",
+    "mochify": "^0.8.2",
+    "react": "^0.11.1",
     "sinon": "^1.9.1",
-    "eslint-jsx": "git+https://github.com/andreypopp/eslint#jsx"
+    "gulp-rename": "^1.2.0",
+    "gulp-react": "^0.4.0",
+    "gulp-clean": "^0.3.1",
+    "gulp-minify-css": "^0.3.6",
+    "gulp-browserify": "^0.5.0",
+    "gulp-less": "^1.3.1",
+    "reactify": "^0.11.0",
+    "gulp": "^3.8.5",
+    "gulp-util": "^2.2.19",
+    "gulp-uglify": "^0.3.1",
+    "gulp-jshint": "^1.6.4",
+    "envify": "^1.2.1",
+    "gulp-connect": "^2.0.6",
+    "gulp-open": "^0.2.8",
+    "glob": "^4.0.4",
+    "gulp-concat": "^2.3.4",
+    "gulp-flatten": "0.0.2",
+    "gulp-plumber": "^0.6.3",
+    "gulp-template": "^0.1.1",
+    "jshint-stylish": "^0.4.0",
+    "gulp-jasmine": "^0.2.0",
+    "underscore": "^1.6.0",
+    "gulp-debug": "^1.0.0",
+    "gulp-add-src": "^0.1.1",
+    "react-tools": "^0.11.1"
+  },
+  "browser": {
+    "react-grid": "./build/js/react-grid.js",
+    "plugin": "./src/javascript/vendor/jquery-plugin.js"
   },
   "browserify": {
     "transform": [
@@ -41,6 +69,10 @@
   },
   "homepage": "https://github.com/prometheusresearch/react-grid",
   "dependencies": {
-    "reactify": "^0.11.0"
+    "reactify": "^0.11.0",
+    "lodash": "^2.4.1",
+    "envify": "^1.2.1",
+    "react": "^0.11.1",
+    "react-tools": "^0.11.1"
   }
 }
