@@ -21,65 +21,81 @@
 
 var React               = require('react');
 var Header              = require('./Header');
+var getWindowSize       = require('./getWindowSize');
 var Viewport            = require('./Viewport');
 var ColumnMetrics       = require('./ColumnMetrics');
-var DOMMetrics          = require('./DOMMetrics');
 
 var Grid = React.createClass({
-    mixins: [ColumnMetrics.Mixin, DOMMetrics.MetricsComputatorMixin],
+    mixins: [ColumnMetrics.Mixin],
 
     propTypes: {
       rows: React.PropTypes.oneOfType([
         React.PropTypes.array.isRequired,
         React.PropTypes.func.isRequired
       ]),
-      rowRenderer: React.PropTypes.component
+      rowRenderer: React.PropTypes.component,
+      length: React.PropTypes.number.isRequired,
+      height: React.PropTypes.number,
+      width: React.PropTypes.number,
+      SelectedCells: React.PropTypes.array,
+      onRowClick: React.PropTypes.func
     },
 
-    style: {
-      overflow: 'hidden',
-      position: 'relative',
-      outline: 0,
-      minHeight: 300
-    },
+
 
     render: function() {
+      var styles = {
+        overflow: 'hidden',
+        position: 'relative',
+        outline: 0,
+        minHeight: 300
+      };
       return this.transferPropsTo(
-        <div style={this.style} className="react-grid-Grid">
+        <div style={styles} className="react-grid-Grid">
           <Header
             ref="header"
             lockedColumns={this.state.lockedColumns}
             regularColumns={this.state.regularColumns}
             onColumnResize={this.onColumnResize}
             height={this.props.rowHeight}
-            totalWidth={this.DOMMetrics.gridWidth()}
+            totalWidth={this.getGridWidth()}
             />
           <Viewport
+            ref="Viewport"
             style={{
-              top: this.props.rowHeight,
               bottom: 0,
               left: 0,
               right: 0,
               position: 'absolute'
             }}
-            width={this.state.lockedColumns.width +
-                this.state.regularColumns.width}
+
+            height={this.props.height - 20}
             rowHeight={this.props.rowHeight}
             rowRenderer={this.props.rowRenderer}
             rows={this.props.rows}
             length={this.props.length}
             lockedColumns={this.state.lockedColumns}
             regularColumns={this.state.regularColumns}
-            totalWidth={this.DOMMetrics.gridWidth()}
+            totalWidth={this.getGridWidth()}
             onViewportScroll={this.onViewportScroll}
+            SelectedCells={this.props.SelectedCells}
+            onRowClick={this.props.onRowClick}
             />
         </div>
       );
     },
+    getGridWidth: function() {
+      var colsWidth = this.state.lockedColumns.width +
+          this.state.regularColumns.width;
+      return colsWidth > this.props.width ? this.props.width : colsWidth;
 
+    },
     getDefaultProps: function() {
+      var winSize = getWindowSize();
       return {
-        rowHeight: 35
+        rowHeight: 35,
+        height: winSize.height,
+        width: winSize.width,
       };
     },
 
