@@ -21,7 +21,6 @@ var React               = require('react/addons');
 var cx                  = React.addons.classSet;
 var utils               = require('./utils');
 var DraggableMixin      = require('./DraggableMixin');
-var getScrollbarSize    = require('./getScrollbarSize');
 var ColumnMetrics       = require('./ColumnMetrics');
 
 var Header = React.createClass({
@@ -29,7 +28,7 @@ var Header = React.createClass({
   propTypes: {
     lockedColumns: React.PropTypes.object.isRequired,
     regularColumns: React.PropTypes.object.isRequired,
-    totalWidth: React.PropTypes.number,
+    totalWidth: React.PropTypes.number.isRequired,
     height: React.PropTypes.number.isRequired
   },
 
@@ -41,13 +40,11 @@ var Header = React.createClass({
       top: 0,
       width: state.lockedColumns.width
     };
-
     var regularColumnsStyle = {
       position: 'absolute',
       top: 0,
-      left: state.lockedColumns.width,
-      width: (this.props.totalWidth -
-              state.lockedColumns.width)
+      left: lockedColumnsStyle.width - 20, //floats over the scrollbar
+      width: (this.props.totalWidth - lockedColumnsStyle.width)
     };
 
     var className = cx({
@@ -163,9 +160,8 @@ var HeaderRow = React.createClass({
   },
 
   render: function() {
-    var scrollbarSize = getScrollbarSize();
     var columnsStyle = {
-      width: this.props.width ? (this.props.width + scrollbarSize) : '100%',
+      width: this.props.width || '100%',
       height: this.props.height,
       position: 'relative',
       whiteSpace: 'nowrap',
@@ -173,7 +169,7 @@ var HeaderRow = React.createClass({
       overflowY: 'hidden'
     };
     return this.transferPropsTo(
-      <div style={this.getStyle()} className="react-grid-HeaderRow">
+      <div style={this.props.style} className="react-grid-HeaderRow">
         <div style={columnsStyle} className="react-grid-HeaderRow__cells">
           {this.props.columns.map((column, idx) => HeaderCell({
             key: idx,
@@ -196,15 +192,6 @@ var HeaderRow = React.createClass({
       || nextProps.columns !== this.props.columns
       || !utils.shallowEqual(nextProps.style, this.props.style)
     );
-  },
-
-  getStyle: function() {
-    return {
-      overflow: 'hidden',
-      width: '100%',
-      height: this.props.height,
-      position: 'relative'
-    };
   }
 
 });
