@@ -18,24 +18,27 @@
  */
 "use strict";
 
+
+/* jshint esnext: true */
+
 var React          = require('react/addons');
 var shallowEqual   = require('./utils').shallowEqual;
-var cx             = React.addons.classSet;
 var cloneWithProps = React.addons.cloneWithProps;
-var ScrollShim     = require('./ScrollShim');
 var Row            = require('./Row');
 
 var Canvas = React.createClass({
-  mixins: [ScrollShim],
 
   propTypes: {
     header: React.PropTypes.component,
     cellRenderer: React.PropTypes.component,
     rowRenderer: React.PropTypes.component,
     rowHeight: React.PropTypes.number.isRequired,
+    height: React.PropTypes.number.isRequired,
     displayStart: React.PropTypes.number.isRequired,
     displayEnd: React.PropTypes.number.isRequired,
     length: React.PropTypes.number.isRequired,
+    style: React.PropTypes.object.isRequired,
+    SelectedCells: React.PropTypes.array.isRequired,
     rows: React.PropTypes.oneOfType([
       React.PropTypes.func.isRequired,
       React.PropTypes.array.isRequired
@@ -54,9 +57,12 @@ var Canvas = React.createClass({
           key: displayStart + idx,
           idx: displayStart + idx,
           row: row,
+          isSelected: this.props.SelectedCells && this.props.SelectedCells[(displayStart + idx)] && this.props.SelectedCells[(displayStart + idx)].length,
+          SelectedCells: this.props.SelectedCells ? this.props.SelectedCells[(displayStart + idx)] : [],
           height: rowHeight,
           columns: this.props.columns,
-          cellRenderer: this.props.cellRenderer
+          cellRenderer: this.props.cellRenderer,
+          onRowClick: this.props.onRowClick
         }));
 
     if (displayStart > 0) {
@@ -70,7 +76,7 @@ var Canvas = React.createClass({
 
     return this.transferPropsTo(
       <div
-        style={{height: this.props.height}}
+        style={this.props.style}
         onScroll={this.onScroll}
         className="react-grid-canvas">
         <div style={{width: this.props.width, overflow: 'hidden'}}>
@@ -82,9 +88,9 @@ var Canvas = React.createClass({
 
   renderRow: function(props) {
     if (React.isValidComponent(this.props.rowRenderer)) {
-      return cloneWithProps(this.props.rowRenderer, props);
+      return cloneWithProps(this.props.rowRenderer(props), props);
     } else {
-      return this.props.rowRenderer(props);
+      return Row(props);
     }
   },
 
@@ -97,10 +103,6 @@ var Canvas = React.createClass({
     );
   },
 
-  getDefaultProps: function() {
-    return {rowRenderer: Row};
-  },
-
   getInitialState: function() {
     return {
       shouldUpdate: true,
@@ -110,8 +112,8 @@ var Canvas = React.createClass({
   },
 
   componentWillReceiveProps: function(nextProps) {
-    var shouldUpdate = !(nextProps.visibleStart > this.state.displayStart
-                        && nextProps.visibleEnd < this.state.displayEnd)
+    var shouldUpdate = !(nextProps.visibleStart >= this.state.displayStart
+                        && nextProps.visibleEnd <= this.state.displayEnd)
                         || nextProps.length !== this.props.length
                         || nextProps.rowHeight !== this.props.rowHeight
                         || nextProps.columns !== this.props.columns
@@ -134,12 +136,10 @@ var Canvas = React.createClass({
   },
 
   onScroll: function(e) {
-    this.appendScrollShim();
     if (this.props.onScroll) {
       this.props.onScroll(e);
     }
   },
-
   setScrollTop: function(scrollTop) {
     this.getDOMNode().scrollTop = scrollTop;
   },
