@@ -4,10 +4,8 @@
 'use strict';
 
 var React = require('react/addons');
-var cx    = React.addons.classSet;
 
 var Cell = React.createClass({
-
   render: function() {
     var style = {
       display: 'block',
@@ -18,8 +16,14 @@ var Cell = React.createClass({
       textOverflow: 'ellipsis',
       overflow: 'hidden'
     };
+    var cx = React.addons.classSet;
+    var classes = cx({
+      'react-grid-Cell': true,
+      'active-cell': this.props.isSelected,
+      'selected-cell': this.props.isSelected
+    });
     return (
-      <div className="react-grid-Cell" style={style}>
+      <div className={classes} tabIndex={this.props.column.idx} style={style} onClick={this.onClick}>
         {this.props.renderer({
           value: this.props.value,
           column: this.props.column
@@ -27,7 +31,9 @@ var Cell = React.createClass({
       </div>
     );
   },
-
+  onClick: function(ev) {
+    if(this.props.onClick) { this.props.onClick(ev, this.props); }
+  },
   getDefaultProps: function() {
     return {
       renderer: simpleCellRenderer
