@@ -3,9 +3,9 @@
  */
 "use strict";
 
+/* jshint esnext: true, unused: false, undef: false */
 var React               = require('react');
 var shallowCloneObject  = require('./shallowCloneObject');
-var DOMMetrics          = require('./DOMMetrics');
 
 /**
  * Update column metrics calculation.
@@ -79,17 +79,14 @@ function resizeColumn(metrics, index, width) {
 }
 
 var Mixin = {
-  mixins: [DOMMetrics.MetricsMixin],
 
   propTypes: {
     columns: React.PropTypes.array,
     minColumnWidth: React.PropTypes.number
   },
 
-  DOMMetrics: {
-    gridWidth: function() {
-      return this.getDOMNode().offsetWidth - 2;
-    }
+  gridWidth: function() {
+    return this.isMounted() ? this.getDOMNode().offsetWidth - 2 : 0;
   },
 
   getDefaultProps: function() {
@@ -107,7 +104,7 @@ var Mixin = {
   },
 
   getColumnMetrics: function(props, initial) {
-    var totalWidth = initial ? null : this.DOMMetrics.gridWidth();
+    var totalWidth = initial ? null : this.gridWidth();
     return {
       regularColumns: calculate({
         columns: props.columns.filter((c) => !c.locked),
