@@ -3,6 +3,8 @@
  */
 'use strict';
 
+/* jshint esnext: true */
+
 var React = require('react/addons');
 var cx    = React.addons.classSet;
 var Cell  = require('./Cell');
@@ -32,18 +34,25 @@ var Row = React.createClass({
     } else {
       children = this.props.columns.map((column, idx) => Cell({
         key: idx,
+        row: this.props.row.key,
+        isSelected: this.props.isSelected && this.props.SelectedCells && this.props.SelectedCells[column.idx],
         value: this.props.row[column.key || idx],
         column: column,
         height: this.props.height,
-        renderer: column.renderer || this.props.cellRenderer
+        renderer: column.renderer || this.props.cellRenderer,
+        onClick: this.onCellClicked
       }));
     }
 
     return this.transferPropsTo(
-      <div className={className} style={style}>
+      <div tabIndex={this.props.row.key} className={className} style={style}>
         {children}
       </div>
     );
+  },
+  onCellClicked: function(ev, cellProps) {
+    //bubble, adding row data
+    if(this.props.onRowClick) { this.props.onRowClick(ev, this.props, cellProps); }
   }
 });
 
