@@ -3,37 +3,28 @@
  *
  * Component hierarchy diagram:
  *
- *  +––––––––––––––––––––––––––––––––––––––––––––––––––––+
- *  | Viewport                                           |
- *  | +––––––––––––––––––––+ +–––––––––––––––––––+ +–––+ |
- *  | | Canvas (locked)    | | Canvas (regular)  | | S | |
- *  | |                    | |                   | | c | |
- *  | |                    | |                   | | r | |
- *  | |                    | |                   | | o | |
- *  | |                    | |                   | | l | |
- *  | |                    | |                   | | l | |
- *  | |                    | |                   | | b | |
- *  | |                    | |                   | | a | |
- *  | |                    | |                   | | r | |
- *  | |                    | +–––––––––––––––––––+ +–––+ |
- *  | |                    | +–––––––––––––––––––––––––+ |
- *  | |                    | | Scrollbar               | |
- *  | +––––––––––––––––––––+ +–––––––––––––––––––––––––+ |
- *  +––––––––––––––––––––––––––––––––––––––––––––––––––––+
+ *  +–––––––––––––––––––––––––––––––––––––––––––––––+
+ *  | Viewport                                      |
+ *  | +––––––––––––––––––––+ +–––––––––––––––––––+  |
+ *  | | Canvas (locked)    | | Canvas (regular)  |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | |                    | |                   |  |
+ *  | +–––––––––––––––––––-+ +–––––––––––––––––––+  |
+ *  +–––––––––––––––––––––––––––––––––––––––––––––––+
  *
  * @jsx React.DOM
  */
 'use strict';
 
 var React             = require('react');
-var Scrollbar         = require('./Scrollbar');
-var getWindowSize     = require('./getWindowSize');
-var getScrollbarSize  = require('./getScrollbarSize');
-var DOMMetrics        = require('./DOMMetrics');
 var Canvas            = require('./Canvas');
-
-var VerticalScrollbar   = Scrollbar.VerticalScrollbar;
-var HorizontalScrollbar = Scrollbar.HorizontalScrollbar;
+var utils            = require('./utils');
 
 var min   = Math.min;
 var max   = Math.max;
@@ -41,16 +32,14 @@ var floor = Math.floor;
 var ceil  = Math.ceil;
 
 var ViewportScroll = {
-  mixins: [DOMMetrics.MetricsMixin],
 
-  DOMMetrics: {
-    viewportHeight: function() {
-      return this.getDOMNode().offsetHeight;
-    }
+  viewportHeight: function() {
+    return this.isMounted() ? this.getDOMNode().offsetHeight : 0;
   },
 
   propTypes: {
     rowHeight: React.PropTypes.number,
+    height: React.PropTypes.number.isRequired,
     length: React.PropTypes.number.isRequired
   },
 
@@ -65,55 +54,45 @@ var ViewportScroll = {
   },
 
   getGridState: function(props) {
-    var height = this.state && this.state.height ?
-      this.state.height :
-      getWindowSize().height;
+    var height = this.props.height;
     var renderedRowsCount = ceil(height / props.rowHeight);
     return {
-      displayStart: 0,
+      displayStart: this.props.initialRow || 0,
       displayEnd: renderedRowsCount * 2,
-      height: height,
-      scrollTop: 0,
-      scrollLeft: 0
+      height: height
     };
   },
 
-  updateScroll: function(scrollTop, scrollLeft, height, rowHeight, length) {
-    var renderedRowsCount = ceil(height / rowHeight);
+  updateScroll: function(height, rowHeight, length, scrollTop, scrollLeft) {
+    scrollTop = scrollTop || 0;
+    scrollLeft = scrollLeft || 0;
+    var nextScrollState = { };
 
-    var visibleStart = floor(scrollTop / rowHeight);
+    nextScrollState.renderedRowsCount = ceil(height / rowHeight);
 
-    var visibleEnd = min(
-        visibleStart + renderedRowsCount,
+    nextScrollState.visibleStart = floor(scrollTop / rowHeight);
+
+    nextScrollState.visibleEnd = min(
+        nextScrollState.visibleStart + nextScrollState.renderedRowsCount,
         length);
 
-    var displayStart = max(
+    nextScrollState.displayStart = max(
         0,
-        visibleStart - renderedRowsCount * 2);
+        nextScrollState.visibleStart - nextScrollState.renderedRowsCount * 2);
 
-    var displayEnd = min(
-        visibleStart + renderedRowsCount * 2,
+    nextScrollState.displayEnd = min(
+        nextScrollState.visibleStart + nextScrollState.renderedRowsCount * 2,
         length);
 
-    var nextScrollState = {
-      visibleStart,
-      visibleEnd,
-      displayStart,
-      displayEnd,
-      height,
-      scrollTop,
-      scrollLeft
-    };
+    nextScrollState.height = height;
 
     this.setState(nextScrollState);
   },
 
   metricsUpdated: function() {
-    var height = this.DOMMetrics.viewportHeight();
+    var height = this.viewportHeight();
     if (height) {
       this.updateScroll(
-        this.state.scrollTop,
-        this.state.scrollLeft,
         height,
         this.props.rowHeight,
         this.props.length
@@ -126,19 +105,35 @@ var ViewportScroll = {
       this.setState(this.getGridState(nextProps));
     } else if (this.props.length !== nextProps.length) {
       this.updateScroll(
-        this.state.scrollTop,
-        this.state.scrollLeft,
-        this.state.height,
+        this.props.height,
         nextProps.rowHeight,
         nextProps.length
       );
     }
-  }
+  },
+
 };
 
 var Viewport = React.createClass({
   mixins: [ViewportScroll],
+  propTypes: {
+    regularColumns: React.PropTypes.object.isRequired,
+    lockedColumns: React.PropTypes.object,
+    onHorizontalScrollUpdate: React.PropTypes.func,
+    onVerticalScrollUpdate: React.PropTypes.func,
+    onViewportScroll: React.PropTypes.func,
+    length: React.PropTypes.number.isRequired,
+    rowHeight: React.PropTypes.number.isRequired,
+    height: React.PropTypes.number.isRequired,
+    rows: React.PropTypes.oneOfType([
+      React.PropTypes.array.isRequired,
+      React.PropTypes.func.isRequired
+    ]),
+    rowRenderer: React.PropTypes.component,
+    SelectedCells: React.PropTypes.array.isRequired,
+    totalWidth: React.PropTypes.number.isRequired
 
+  },
   style: {
     overflowX: 'hidden',
     overflowY: 'hidden',
@@ -147,28 +142,16 @@ var Viewport = React.createClass({
   },
 
   render: function() {
-    var shift = getScrollbarSize();
     var locked = this.renderLockedCanvas();
     var regular = this.renderRegularCanvas();
+    this.style.height = this.props.height;
     return this.transferPropsTo(
       <div
         className="react-grid-Viewport"
         style={this.style}>
         {locked && locked.canvas}
         {regular.canvas}
-        {shift > 0 && <HorizontalScrollbar
-          size={regular.style.width - shift}
-          totalSize={this.props.regularColumns.width}
-          style={{left: regular.style.left}}
-          position={this.state.scrollLeft}
-          onScrollUpdate={this.onHorizontalScrollUpdate}
-          />}
-        {shift > 0 && <VerticalScrollbar
-          size={this.state.height}
-          totalSize={this.props.length * this.props.rowHeight}
-          position={this.state.scrollTop}
-          onScrollUpdate={this.onVerticalScrollUpdate}
-          />}
+
       </div>
     );
   },
@@ -177,70 +160,44 @@ var Viewport = React.createClass({
     if (this.props.lockedColumns.columns.length === 0) {
       return null;
     }
-
-    var shift = getScrollbarSize();
-    var width = this.props.lockedColumns.width + shift;
-    var hScroll = this.props.lockedColumns.width > width;
-
-    var style = {
-      position: 'absolute',
-      top: 0,
-      width: width,
-      overflowX: hScroll ? 'scroll' : 'hidden',
-      overflowY: 'scroll',
-      paddingBottom: hScroll ? shift : 0
-    };
-
-    var canvas = (
-      <Canvas
-        ref="lockedRows"
-        className="react-grid-Viewport__locked"
-        style={style}
-        width={this.props.lockedColumns.width}
-        rows={this.props.rows}
-        columns={this.props.lockedColumns.columns}
-        rowRenderer={this.props.rowRenderer}
-
-        visibleStart={this.state.visibleStart}
-        visibleEnd={this.state.visibleEnd}
-        displayStart={this.state.displayStart}
-        displayEnd={this.state.displayEnd}
-
-        length={this.props.length}
-        height={this.state.height + (hScroll ? shift : 0)}
-        rowHeight={this.props.rowHeight}
-        onScroll={this.onScroll.bind(null, "lockedRows")}
-        />
-    );
-    return {canvas, style};
+    var width = this.props.lockedColumns.width;
+    return this.renderAnyCanvas("locked", this.props.lockedColumns,{width:width});
   },
 
   renderRegularCanvas: function() {
-    var shift = getScrollbarSize();
-    var width = (this.props.totalWidth -
-                 this.props.lockedColumns.width +
-                 shift);
-    var hScroll = this.props.regularColumns.width > width;
-
-    var style = {
+    var width = (this.props.totalWidth - this.props.lockedColumns.width);
+    return this.renderAnyCanvas("regular", this.props.regularColumns,{
+      width:width,
+      left:this.props.lockedColumns.width - 20,//floats over the scrollbar
+      backgroundColor:'white' //and 'hides' it
+      });
+  },
+  componentWillMount: function() {
+    this._ignoreNextScroll = this._ignoreNextScroll || null;
+  },
+  renderAnyCanvas: function(canvasId, cols, style) {
+    var hScroll = cols.width > style.width;
+    utils.mergeInto(style, {
       position: 'absolute',
       top: 0,
       overflowX: hScroll ? 'scroll' : 'hidden',
       overflowY: 'scroll',
-      width: width,
-      left: this.props.lockedColumns.width,
-      paddingBottom: hScroll ? shift : 0
-    };
+      height: this.props.height - 20,
+      marginBottom: hScroll ? 20 : 0
+    });
 
+    var canvasRef=canvasId + "Rows";
+    var canvasClass="react-grid-Viewport__" + canvasId;
     var canvas = (
       <Canvas
-        ref="regularRows"
-        className="react-grid-Viewport__regular"
-        width={this.props.regularColumns.width}
+        ref={canvasRef}
+        className={canvasClass}
         style={style}
+        width={cols.width}
         rows={this.props.rows}
-        columns={this.props.regularColumns.columns}
+        columns={cols.columns}
         rowRenderer={this.props.rowRenderer}
+        SelectedCells={this.props.SelectedCells}
 
         visibleStart={this.state.visibleStart}
         visibleEnd={this.state.visibleEnd}
@@ -248,15 +205,13 @@ var Viewport = React.createClass({
         displayEnd={this.state.displayEnd}
 
         length={this.props.length}
-        height={this.state.height + (hScroll ? shift : 0)}
         rowHeight={this.props.rowHeight}
-        onScroll={this.onScroll.bind(null, "regularRows")}
+        onScroll={this.onScroll.bind(null, {canvasRef})}
+        onRowClick={this.props.onRowClick}
         />
     );
-
     return {canvas, style};
   },
-
   onScroll: function(rowGroup, e) {
     if (this._ignoreNextScroll !== null &&
         this._ignoreNextScroll !== rowGroup) {
@@ -267,10 +222,9 @@ var Viewport = React.createClass({
     // we do this outside of React for better performance...
     // XXX: we might want to use rAF here
     var scrollTop = e.target.scrollTop;
-    var scrollLeft = rowGroup === 'lockedRows' ?
-      this.state.scrollLeft : e.target.scrollLeft;
+    var scrollLeft = e.target.scrollLeft;
 
-    var toUpdate = rowGroup === 'lockedRows' ?
+    var toUpdate = rowGroup.canvasRef === 'lockedRows' ?
         this.refs.regularRows :
         this.refs.lockedRows;
 
@@ -280,11 +234,11 @@ var Viewport = React.createClass({
     }
 
     this.updateScroll(
-      scrollTop,
-      scrollLeft,
-      this.state.height,
+      this.props.height,
       this.props.rowHeight,
-      this.props.length
+      this.props.length,
+      scrollTop,
+      scrollLeft
     );
 
     if (this.props.onViewportScroll) {
@@ -292,13 +246,7 @@ var Viewport = React.createClass({
     }
   },
 
-  onVerticalScrollUpdate: function(scrollTop) {
-    this.refs.regularRows.getDOMNode().scrollTop = scrollTop;
-  },
 
-  onHorizontalScrollUpdate: function(scrollLeft) {
-    this.refs.regularRows.getDOMNode().scrollLeft = scrollLeft;
-  }
 });
 
 module.exports = Viewport;
